+++ Headers/ConsoleFun_GotoRC.h
@@ -8,30 +8,31 @@
 //                                                                     //
 // See the License for the permissions and limitations.                //
 
-#ifdef __cplusplus
-extern "C"
+#ifdef ConsoleFun_CFOpts_
+#define ConsoleFun_GotoRC_ 1
+
+void ConsoleFun_GotoRC(int Row, int Col)
 {
-#endif
-
-#ifdef ConsoleFun_Include
-
-    void ConsoleFun_GotoRC(int Row, int Col)
-    {
 #ifdef __linux__
-        printf("%c[%d;%df", 0x1B, Row + 1, Col + 1);
+    printf("%c[%d;%df", 0x1B, Row + 1, Col + 1);
 #elif _WIN32
-        HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE);
-        COORD Coordinates;
+    HANDLE hConsole;
+    COORD Coordinates;
+    if (ConsoleFun.Cache.hConsoleOut == NULL)
+    {
+        hConsole = GetStdHandle(STD_OUTPUT_HANDLE);
+        ConsoleFun.Cache.hConsoleOut = hConsole;
+    }
+    else
+    {
+        hConsole = ConsoleFun.Cache.hConsoleOut;
+    }
 
-        Coordinates.X = Col;
-        Coordinates.Y = Row;
-        SetConsoleCursorPosition(hConsole, Coordinates);
+    Coordinates.X = Col;
+    Coordinates.Y = Row;
+    SetConsoleCursorPosition(hConsole, Coordinates);
 #endif
-        return;
-    } // ConsoleFun_GotoRC
+    return;
+} // ConsoleFun_GotoRC
 
-#endif // ConsoleFun_Include
-
-#ifdef __cplusplus
-}
-#endif
+#endif // ConsoleFun_CFOpts_